% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_styles.R
\name{nc_color_styles}
\alias{nc_color_styles}
\title{Available coordinated figure color styles}
\usage{
nc_color_styles()
}
\value{
Table of style names and their categorical, sequential and diverging bases.
}
\description{
Available coordinated figure color styles
}
